Sensor Data Superbase Upload/requirements.txt
requests
pyserial
dotenv